Revert "replace time measurement method"
This reverts commit 628c595db3f3d06e381c67a9cfd51933877fd315.

+++ SilverSim/Viewer.Core/SimCircuit.cs
@@ -222,7 +222,7 @@ namespace SilverSim.Viewer.Core
             /* no additional threads required */
         }
 
-        protected override void SendSimStats(long dt)
+        protected override void SendSimStats(int dt)
         {
             /* no sim stats */
         }
